[prim_onehot_check] Add option for permissive en_i checks
A more permissive en_i check is needed when for example the oh_i vector
is partially masked. E.g. when checking CSR WE strobes, it can happen
that some of the WE strobes are additionally masked by REGWEN bits.
In that case there may be situations where the en_i (connected to the
global write enable bit) is 1, but all  of the oh_i (connected to the
individual register we bits) elements are 0.

Signed-off-by: Michael Schaffner <msf@google.com>