[dv/xcelium] Fix Xcelium nightly regression error
This PR fixes the xcelium nightly regression error because we export
`cov_merge_cmd` in the enviornment.
In the command line, current what we see is that:
cov_merge_cmd={prefix} imc
The space between {prefix} and imc will cause trouble and
`cov_merge_cmd` ends up not loading the correct cmd.

To fix this issue, we add a single quotation, now the export command
looks like this:
cov_merge_cmd=`{prefix} imc`

Thanks @weicaiyang for debugging this issue.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>