[bazel] update `build_otbn.py` to import otbn_ld module
The `build_otbn.py` was invoking the OTBN linker Python script via
the shell, instead of importing the module directly. This does not
enable bazel to make use of the hermetic python toolchain and packages.
This updates the `build_otbn.py` to invoke the OTBN linker Python
script via a direct function call from a module import.

Signed-off-by: Timothy Trippel <ttrippel@google.com>