[dv/cdc assertion] Temp remove CDC assertion cov collection in VCS
This PR temp remove cdc DV module assertion cov in regression. Right now
this temp removal only on VCS because adc_ctrl seems to be the only IP
blocked by this to reach V2.
Once CDC is fully supported in DV simulation, we can remove this. So I
did not implement it in Xcelium.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>