[bazel] Delete sw/vendor/riscv_compliance
Unlike most code in sw/vendor, it is not used by either the Meson or
Bazel builds directly, and is instead invoked through their Make files.
Because Bazel knows how to independently build this repository, and,
unlike the Make version, run the tests, there is no reason to keep it or
its associated CI steps.

In particular, because CI already builds every test, we were *already*
doing what the CI step used to do as part of it.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>