[sw,tests] Test flash_ctrl init and scramble
For tests:
chip_sw_flash_init
chip_sw_flash_scramble

Tests the initialization and scramble functions of the flash_ctrl. Checks that data written to unscrambled
partitions is still accessible following the initialization. Also checks that data written to scrambled partitons
in no longer accessible following initialization when new keys are fetched from OTP.
Checks that the initialization process writes seeds to the keymgr when lc_seed_hw_rd_en is set and does
not when it is unset. Does backdoor writes of scrambled data to the flash to test the flash_ctrl unscrambling.

Signed-off-by: Dave Williams <dave.williams@ensilica.com>