[sw/benchmarks] fix coremark benchmark nightly test
The coremark benchmark test meson build target was commented out in
PR #12206, since it failed to build when the OTTF entry point was
renamed from `main` to `_ottf_main` (the `ottf_start.asm` assembly is
used with the benchmark code to setup the device). However, this benchmark
is run as a nightly regression in DV simulation. As a result of the missing
build target, this test was failing.

This patches the coremark benchmark code to work with the new entry
point symbol: `_ottf_main`.

Signed-off-by: Timothy Trippel <ttrippel@google.com>