fix(spi_device): Re-timing sck_cmdfifo_set clear time
Related issue #11871

In previous design, the `sck_cmdfifo_set` signal is set by the ACK of
CMDFIFO write port and cleared by the FIFO DEPTH == 0.

THe clearing of FIFO depth depends on the SW processing time. This
commit revises it to CSb assertion pulse signal.

It is still bounded by the clock ratio between the SYS_CLK and the SCK.
SYS_CLK should not be 3 times slower than the SCK to correctly generate
the pulse signal within the opcode timeline.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>