[fpga, doc, util] Remove refs and utils related to NexysVideo FPGA board
The ChipWhisperer CW310 board is the main FPGA development board for
OpenTitan. It has been decided to drop support for the NexysVideo board,
see lowRISC/OpenTitan#7814. Thus, this commit removes references to
the NexysVideo board from the documentation as well as tooling and
utilities specific to the NexysVideo board.

This is related to lowRISC/OpenTitan#12221.
This resolves lowRISC/OpenTitan#12185.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>