[bazel] Fix bazel build of englishbreakfast
The bazel EB build needs to reference the correct autogen-ed files. In
meson, this was accomplished by replacing the TOPNAME variable. For
bazel, we take a different approach. Instead of modifying the BUILD
files, we instead move the autogen-ed files from the EB directories to
the Earlgrey directories where bazel is already pointed.

Signed-off-by: Miles Dai <milesdai@google.com>