[dif/lc_ctrl] split `dif_lc_ctrl_transition` into two
As identified during the lc_ctrl D3 review, it is useful to be able to
give the programmer a chance to check the lifecycle transition
state/token/clock-enable registers _before_ issuing a transition
command, as there are a limited number of lifecycle transitions that can
be performed on each chip.

This fixes #12822.

Signed-off-by: Timothy Trippel <ttrippel@google.com>