[otbn,dv] Simplify how we track memory wiping in ISS
This slight rejig gets rid of some repeated code but, more
importantly, also sorts out how we report changes when we're in the
MEM_SEC_WIPE fsm state.

The fact that we weren't reporting them properly caused occasional
problems because we couldn't drop RND_REQ if a prefetch completed in
the middle of a memory wipe.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>