[rstmgr] Address comment from #12890
The consistency check default state does not currently signal an error,
it creates a situation where a continuous glitch or a
severed enable/clock to the fsm flops could suppress the error/alert.

This does not seem feasible, but it's better to be safe since it
has a very low cost.

Signed-off-by: Timothy Chen <timothytim@google.com>