[top] Change usbdev bus clock to usb clock
For top_earlgrey, also move usbdev to xbar_main. This keeps the number
of clock domain crossings down and should improve latency considerably.
In a follow-up, much of usbdev's CDC for CSRs will be removed.

Because nearly the entire IP is clocked by the 48 MHz USB clock and it
is of a similar magnitude as the main clock, it's best to place the CDC
in the crossbar instead of at individual register sites.

Signed-off-by: Alexander Williams <awill@google.com>