[regtool] Add option to scrub alias definitions
This adds the option to generate generic register definitions from a
an alias register definition. It can be used as follows:

regtool.py <path to generic hjson> --alias <path to alias hjson> --scrub -j

This will replace ALL CSR descriptions in the aliased interface and
replace them with a scrubbed version of the alias definitions.

Signed-off-by: Michael Schaffner <msf@opentitan.org>