[sw,tests] Update SRAM scramble tests to handle ECC errors
For tests:
chip_sw_sram_ctrl_main_scrambled_access
chip_sw_sram_ctrl_main_scrambled_access_jitter_en
chip_sw_sram_ctrl_ret_scrambled_access

The tests have been updated to handle ECC errors in the readback data. Previously,
this was not needed as ECC checking was not enabled. It was sufficient to write
data, scramble, read back and check for mismatches against the original data. ECC errors
now trigger an internal interrupt in the CPU. An exception handler has been written for
this. In addition to the read back check, a count of the number of exceptions handled
is kept and checked against the expected number.

Signed-off-by: Dave Williams <dave.williams@ensilica.com>