[opentitantool]: Adapt heuristics for Ti50 images
As part of the "rescue" protocol for Ti50 updates, OpenTitan tool needs
to recognize the "RW" part of the given image file.  This is done
through heuristics that recognize the header of each section, and
relying on the order in which sections appear.

Recent images happen to have the header signature in more places in the
image, causing mis-identification of the RW section.  This CL improves
the heuristics.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I3f9e0003ecf71b3967d25af88db962b87d041b0f