[bazel] clean up failing_verilator and cpu:4 tags
The "failing_verilator" tag was doing the same thing as the "broken" tag and
required bazel users to remember to filter for both, so all instances of
"failing_verilator" are replaced by broken.

"cpu:4" is applied to all verilator tests by rules/opentitan_test.bzl so
other applications are removed and it's removed from the cw310 tests
which are exclusive and aren't affected by it.

Signed-off-by: Drew Macrae <drewmacrae@google.com>